- successfully joined a MOHAA server. there's still a lot that can go wrong though - especially CG messages *sigh*

+++ code/qcommon/q_shared.h
@@ -1125,8 +1125,8 @@ typedef struct playerState_s {
 typedef struct usercmd_s {
 	int				serverTime;
 	byte msec;
-	short unsigned int buttons;
-	short int		angles[3];
+	int buttons;
+	int		angles[3];
 //	int 			buttons;
 	byte			weapon;           // weapon
 	signed char	forwardmove, rightmove, upmove;
